requirements.txt
# Autonomous Research Agent
# Dependencies for LangGraph-based multi-model research agent

# Core LangGraph/LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0

# HTTP and API clients
httpx>=0.27.0
aiohttp>=3.9.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Output and logging
rich>=13.0.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Frontend
streamlit>=1.40.0
plotly>=5.18.0
